Updated plot mfcc and spectrogram features
Previously using the axes divider did not separate the colobar from
the mfcc or spectrogram plots. On Windows 10, I found that the divider
caused the colobar to take over the entire plot. I knew they were
trying to have the colorbar on the rightside of the plot. So,
I passed ax in list to colorbar and location as right to colobar,
then the colorbar appeared on the rightside of the plot without taking
over the plot completely.